requirements-dev.txt
# Development and testing dependencies
-r requirements.txt

# Testing framework
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pytest-mock>=3.14.0

# Code quality
black>=24.10.0
ruff>=0.7.0
mypy>=1.13.0
pre-commit>=4.0.0

# Type stubs
pandas-stubs>=2.2.3
types-aiofiles>=24.1.0
types-tabulate>=0.9.0
types-pytz>=2024.2

# Documentation
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.26.0

# Development tools
ipython>=8.29.0
jupyter>=1.1.0
rich>=13.9.0  # Better console output